backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic[email]==2.10.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.2.0
httpx==0.28.0
beautifulsoup4==4.12.3
celery==5.4.0
python-dotenv==1.0.1
alembic==1.14.0
apscheduler==3.10.4
playwright
playwright-stealth==2.0.2
